refactor: refactor the enum types
